advances

+++ src/controllers/cart.controllers.js
@@ -41,14 +41,30 @@ const create = catchError(async(req, res) => {
 
 const getOne = catchError(async(req, res) => {
     const { id } = req.params;
-    const result = await Cart.findByPk(id);
+    const userId = req.user.id
+    const result = await Cart.findByPk(id,{
+        where:{ userId: userId },
+        include:[
+            {
+                model:Product,
+                attributes:{exclude:['createdAt','updatedAt']},
+                include:[
+                    {
+                     model:Category,
+                     attributes: ['name']
+                    }
+                ]
+            }
+        ]
+    });
     if(!result) return res.sendStatus(404);
     return res.json(result);
 });
 
 const remove = catchError(async(req, res) => {
     const { id } = req.params;
-    const result = await Cart.destroy({ where: {id} });
+    const userId = req.user.id
+    const result = await Cart.destroy({ where: {id,userId} });
     if(!result) return res.sendStatus(404);
     return res.sendStatus(204);
 });
@@ -61,11 +77,12 @@ const update = catchError(async(req, res) => {
 
     const result = await Cart.update(
         req.body,
-        { where: {id}, returning: true }
+        { where: {id, userId: req.user.id}, returning: true }
     );
     if(result[0] === 0) return res.sendStatus(404);
     return res.json(result[1][0]);
 });
+
 
 module.exports = {
     getAll,
